template divided
