+++ system/views/__init__.py
@@ -1,1 +1,4 @@
-from .views import *
+from .home import *
+from .overview import *
+from .profile import *
+from .register import *
